wactest/requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow
django-extensions
django-extra-fields
mock